test: 예외처리 확인
