+++ src/test/java/racingcar/RacingcarServiceTest.java
@@ -17,29 +17,9 @@ class RacingcarServiceTest {
     private RacingcarService racingcarService = new RacingcarService();
 
     @Test
-    @DisplayName("구분자가 없을 때 예외")
-    void validSeparator() {
-        String input = "우가.밀리";
-
-        assertThatThrownBy(() -> racingcarService.makeCars(input))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 구분자(,)가 필요해요.");
-    }
-
-    @Test
-    @DisplayName("경주 참가자가 한명이하일 때 예외")
-    void validParticipants() {
-        String input = "밀리,";
-
-        assertThatThrownBy(() -> racingcarService.makeCars(input))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 경주는 최소 2명이 필요해요.");
-    }
-
-    @Test
     @DisplayName("입력값을 받아서 자동차 리스트 생성 확인")
     void makeCars() {
-        String input = "밀리,우가";
+        String[] input = {"밀리", "우가"};
 
         Cars cars = racingcarService.makeCars(input);
         assertAll(
